fix: move xqueue services
