+++ xmodule/services.py
@@ -6,11 +6,14 @@ Module contains various XModule/XBlock services
 import inspect
 import logging
 from functools import partial
+from typing import TYPE_CHECKING
 
 from config_models.models import ConfigurationModel
 from django.conf import settings
+from django.urls import reverse
 from eventtracking import tracker
 from edx_when.field_data import DateLookupFieldData
+from requests.auth import HTTPBasicAuth
 from xblock.reference.plugins import Service
 from xblock.runtime import KvsFieldData
 
@@ -22,8 +25,13 @@ from lms.djangoapps.courseware.field_overrides import OverrideFieldData
 from lms.djangoapps.courseware.model_data import DjangoKeyValueStore, FieldDataCache
 from lms.djangoapps.lms_xblock.field_data import LmsFieldData
 from lms.djangoapps.lms_xblock.models import XBlockAsidesConfig
+from openedx.core.djangoapps.waffle_utils import CourseWaffleFlag
+from xmodule.capa.xqueue_interface import XQueueInterface
 
 from lms.djangoapps.grades.api import signals as grades_signals
+
+if TYPE_CHECKING:
+    from xmodule.capa_block import ProblemBlock
 
 log = logging.getLogger(__name__)
 
@@ -308,3 +316,98 @@ class EventPublishingService(Service):
         # in order to avoid duplicate work and possibly conflicting semantics.
         if not getattr(block, 'has_custom_completion', False):
             self.completion_service.submit_completion(block.scope_ids.usage_id, 1.0)
+
+
+# .. toggle_name: send_to_submission_course.enable
+# .. toggle_implementation: CourseWaffleFlag
+# .. toggle_description: Enables use of the submissions service instead of legacy xqueue for course problem submissions.
+# .. toggle_default: False
+# .. toggle_use_cases: opt_in
+# .. toggle_creation_date: 2024-04-03
+# .. toggle_expiration_date: 2025-08-12
+# .. toggle_will_remain_in_codebase: True
+# .. toggle_tickets: none
+# .. toggle_status: supported
+SEND_TO_SUBMISSION_COURSE_FLAG = CourseWaffleFlag("send_to_submission_course.enable", __name__)
+
+
+class XQueueService:
+    """
+    XBlock service providing an interface to the XQueue service.
+
+    Args:
+        block: The `ProblemBlock` instance.
+    """
+
+    def __init__(self, block: "ProblemBlock"):
+        self._block = block
+
+        basic_auth = settings.XQUEUE_INTERFACE.get("basic_auth")
+        requests_auth = HTTPBasicAuth(*basic_auth) if basic_auth else None
+
+        use_submission = self.use_edx_submissions_for_xqueue()
+
+        self._interface = XQueueInterface(
+            settings.XQUEUE_INTERFACE["url"],
+            settings.XQUEUE_INTERFACE["django_auth"],
+            requests_auth,
+            block=block,
+            use_submission_service=use_submission,
+        )
+
+    @property
+    def interface(self):
+        """
+        Returns the XQueueInterface instance.
+        """
+        return self._interface
+
+    def use_edx_submissions_for_xqueue(self) -> bool:
+        """
+        Determines whether edx-submissions should be used instead of legacy XQueue.
+
+        This helper abstracts the toggle logic so that the rest of the codebase is not tied
+        to specific feature flag mechanics or rollout strategies.
+
+        Returns:
+            bool: True if edx-submissions should be used, False otherwise.
+        """
+        return SEND_TO_SUBMISSION_COURSE_FLAG.is_enabled(self._block.scope_ids.usage_id.context_key)
+
+    def construct_callback(self, dispatch: str = "score_update") -> str:
+        """
+        Return a fully qualified callback URL for the external queueing system.
+        """
+        course_key = self._block.scope_ids.usage_id.context_key
+        userid = str(self._block.scope_ids.user_id)
+        mod_id = str(self._block.scope_ids.usage_id)
+
+        callback_type = "xqueue_callback"
+
+        relative_xqueue_callback_url = reverse(
+            callback_type,
+            kwargs={
+                "course_id": str(course_key),
+                "userid": userid,
+                "mod_id": mod_id,
+                "dispatch": dispatch,
+            },
+        )
+
+        xqueue_callback_url_prefix = settings.XQUEUE_INTERFACE.get("callback_url", settings.LMS_ROOT_URL)
+        return f"{xqueue_callback_url_prefix}{relative_xqueue_callback_url}"
+
+    @property
+    def default_queuename(self) -> str:
+        """
+        Returns the default queue name for the current course.
+        """
+        course_id = self._block.scope_ids.usage_id.context_key
+        return f"{course_id.org}-{course_id.course}".replace(" ", "_")
+
+    @property
+    def waittime(self) -> int:
+        """
+        Returns the number of seconds to wait in between calls to XQueue.
+        """
+        return settings.XQUEUE_WAITTIME_BETWEEN_REQUESTS
