+++ xmodule/tests/test_services.py
@@ -2,20 +2,22 @@
 Tests for SettingsService
 """
 import unittest
-from unittest import mock
+from unittest import TestCase, mock
+from unittest.mock import Mock, patch
 
-import pytest
-from django.test import TestCase
 import ddt
-
+import pytest
 from config_models.models import ConfigurationModel
 from django.conf import settings
 from django.test.utils import override_settings
+from opaque_keys.edx.locator import BlockUsageLocator, CourseLocator
+from xblock.fields import ScopeIds
 from xblock.runtime import Mixologist
 
-from opaque_keys.edx.locator import CourseLocator
-from xmodule.services import ConfigurationService, SettingsService, TeamsConfigurationService
+from openedx.core.djangolib.testing.utils import skip_unless_lms
 from openedx.core.lib.teams_config import TeamsConfig
+from xmodule.capa.xqueue_interface import XQueueInterface
+from xmodule.services import ConfigurationService, SettingsService, TeamsConfigurationService, XQueueService
 
 
 class _DummyBlock:
@@ -163,3 +165,71 @@ class TestTeamsConfigurationService(ConfigurationServiceBaseClass):
     def test_get_teamsconfiguration(self):
         teams_config = self.configuration_service.get_teams_configuration(self.course.id)
         assert teams_config == self.teams_config
+
+
+@pytest.mark.django_db
+@skip_unless_lms
+class XQueueServiceTest(TestCase):
+    """Test the XQueue service methods."""
+
+    def setUp(self):
+        super().setUp()
+        location = BlockUsageLocator(
+            CourseLocator("test_org", "test_course", "test_run"),
+            "problem",
+            "ExampleProblem",
+        )
+        self.block = Mock(scope_ids=ScopeIds("user1", "mock_problem", location, location))
+        self.block.max_score = Mock(return_value=10)  # Mock max_score method
+        self.service = XQueueService(self.block)
+
+    def test_interface(self):
+        """Test that the `XQUEUE_INTERFACE` settings are passed from the service to the interface."""
+        assert isinstance(self.service.interface, XQueueInterface)
+        assert self.service.interface.url == "http://sandbox-xqueue.edx.org"
+        assert self.service.interface.auth["username"] == "lms"
+        assert self.service.interface.auth["password"] == "***REMOVED***"
+        assert self.service.interface.session.auth.username == "anant"
+        assert self.service.interface.session.auth.password == "agarwal"
+
+    @patch("lms.djangoapps.grades.services.XQueueService.use_edx_submissions_for_xqueue", return_value=True)
+    def test_construct_callback_with_flag_enabled(self, mock_flag):  # pylint: disable=unused-argument
+        """Test construct_callback when the waffle flag is enabled."""
+        self.service = XQueueService(self.block)
+        usage_id = self.block.scope_ids.usage_id
+        course_id = str(usage_id.course_key)
+        callback_url = f"courses/{course_id}/xqueue/user1/{usage_id}"
+
+        assert self.service.construct_callback() == f"{settings.LMS_ROOT_URL}/{callback_url}/score_update"
+        assert self.service.construct_callback("alt_dispatch") == (
+            f"{settings.LMS_ROOT_URL}/{callback_url}/alt_dispatch"
+        )
+
+        custom_callback_url = "http://alt.url"
+        with override_settings(XQUEUE_INTERFACE={**settings.XQUEUE_INTERFACE, "callback_url": custom_callback_url}):
+            assert self.service.construct_callback() == f"{custom_callback_url}/{callback_url}/score_update"
+
+    @patch("lms.djangoapps.grades.services.XQueueService.use_edx_submissions_for_xqueue", return_value=False)
+    def test_construct_callback_with_flag_disabled(self, mock_flag):  # pylint: disable=unused-argument
+        """Test construct_callback when the waffle flag is disabled."""
+        self.service = XQueueService(self.block)
+        usage_id = self.block.scope_ids.usage_id
+        callback_url = f"courses/{usage_id.context_key}/xqueue/user1/{usage_id}"
+
+        assert self.service.construct_callback() == f"{settings.LMS_ROOT_URL}/{callback_url}/score_update"
+        assert self.service.construct_callback("alt_dispatch") == f"{settings.LMS_ROOT_URL}/{callback_url}/alt_dispatch"
+
+        custom_callback_url = "http://alt.url"
+        with override_settings(XQUEUE_INTERFACE={**settings.XQUEUE_INTERFACE, "callback_url": custom_callback_url}):
+            assert self.service.construct_callback() == f"{custom_callback_url}/{callback_url}/score_update"
+
+    def test_default_queuename(self):
+        """Check the format of the default queue name."""
+        assert self.service.default_queuename == "test_org-test_course"
+
+    def test_waittime(self):
+        """Check that the time between requests is retrieved correctly from the settings."""
+        assert self.service.waittime == 5
+
+        with override_settings(XQUEUE_WAITTIME_BETWEEN_REQUESTS=15):
+            assert self.service.waittime == 15
